Remove cfg-if and move mobile in its own file
