+++ nym-vpn-core/Cargo.toml
@@ -96,7 +96,6 @@ bs58 = "0.5.1"
 bytes = "1.11.1"
 cc = "1.2.62"
 celes = "2.8.2"
-cfg-if = "1.0.4"
 colored = "3.1.1"
 clap = "4.6"
 dbus = "0.9.11"
